+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added auto-save feature when writing a new post.
 - Added privacy and terms screens.
 - Improved some screen transitions.
+- Added automatic conversion of web image and video URLs when writing a post.
 - Updated a dependency to the latest release.
 
 ## Version 0.1.3 (July 22, 2024)